+++ src/pages/index.tsx
@@ -42,6 +42,7 @@ const Home: NextPage = () => {
   if (user) {
     user.preferences = session?.preferences as Preferences;
   }
+
   const groupedRides = groupRides(data);
   const ridesFound = groupedRides.length > 0;